Move: 공통 타입 파일 이동 ('types/common')
