+++ src/types/common/datetime.ts
@@ -1,4 +1,4 @@
-import { CALENDAR_FORMS } from '../data/constant';
+import { CALENDAR_FORMS } from '../../data/constant';
 
 export type CalendarForm = (typeof CALENDAR_FORMS)[number];
 export type Months = 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12;
